Add comprehensive test suite for PracticeQueueBuilder
Testing Infrastructure:
- Extract repository interfaces (INounRepository, IVerbRepository, IUserDataRepository)
- Add ILemmaRepository base interface for shared operations
- Create fake implementations for isolated unit testing
- Add TestScenarioBuilder for fluent test scenario setup

Test Coverage (43 new tests):
- Determinism: same seed produces same queue, different seeds differ
- Bucket rotation: round-robin through mastery level buckets
- Slot spacing: new forms appear every 4-6 reviews
- Spacing constraints: lemma, combo, category gaps with fallback chain
- EffectiveGap scaling: adapts to pool diversity (1-20 lemmas)
- Pool exhaustion: graceful handling when pools run out
- SRS invariants: cooldown exclusion, retired forms (level 11) excluded
- Eligibility: attestation filtering, lemma range boundaries
- Verb rules: reflexive exclusion, pr_3sg skipped
- Settings canonicalization: order-independent determinism

Algorithm Improvements:
- Add tie-breakers for deterministic ordering (FormId, LemmaId)
- Canonicalize settings lists for consistent iteration order
- Change Declension.ParseId to accept long for consistency

+++ PaliPractice/PaliPractice/Models/Inflection/Declension.cs
@@ -69,14 +69,14 @@ public class Declension
     /// <summary>
     /// Parses a FormId back into its component parts.
     /// </summary>
-    public static (int LemmaId, Case Case, Gender Gender, Number Number, int EndingId) ParseId(int formId)
+    public static (int LemmaId, Case Case, Gender Gender, Number Number, int EndingId) ParseId(long formId)
     {
         return (
-            formId / LemmaDivisor,
+            (int)(formId / LemmaDivisor),
             (Case)(formId % LemmaDivisor / 1_000),
             (Gender)(formId % 1_000 / 100),
             (Number)(formId % 100 / 10),
-            formId % 10
+            (int)(formId % 10)
         );
     }
